+++ core/src/kdynobjs.c
@@ -4,7 +4,7 @@
 /** RK0 - The Embedded Real-Time Kernel '0'                                   */
 /** (C) 2026 Antonio Giacomelli <dev@kernel0.org>                             */
 /**                                                                           */
-/** VERSION: V0.73.0                                                          */
+/** VERSION: V0.73.1                                                          */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
 /** http://www.apache.org/licenses/LICENSE-2.0                                */
@@ -861,7 +861,8 @@ RK_ERR kMRMDestroy(RK_MRM_HANDLE *const mrmHandlePtr)
 
     ULONG const currentHeld = (mrmPtr->currBufPtr != NULL) ? 1UL : 0UL;
     if (((mrmPtr->currBufPtr != NULL) &&
-         (mrmPtr->currBufPtr->nUsers > 0UL)) ||
+         ((mrmPtr->currBufPtr->nUsers > 0UL) ||
+          (mrmPtr->currBufPtr->reserved == RK_TRUE))) ||
         ((mrmPtr->mrmMem.nMaxBlocks - mrmPtr->mrmMem.nFreeBlocks) !=
          currentHeld) ||
         ((mrmPtr->mrmDataMem.nMaxBlocks - mrmPtr->mrmDataMem.nFreeBlocks) !=